Moved icons to Editor Default Resources and replaced GUISkin with one defined in code
Everything that’s put in the Resources folders are included in the build. Editor resources are useless in builds. That’s why I moved the textures to the Editor’s own Resources folder.

Having a complete GUISkin asset just for defining width and margin is, in my opinion, a bit too much. There are lots of settings in the GUISkin asset, and without knowing how the ExtendedTransformEditor works, it’s hard to know which settings are important. By defining the GUIStyle in code, it’s much more clear which settings matters.